Removed bootstrap command. Pull does not clone missing repos anymore. Update was renamed to sync.
