+++ lib/commands/exec.js
@@ -47,7 +47,7 @@ module.exports = {
 
 			// Package does not exist.
 			if ( !fs.existsSync( newCwd ) ) {
-				log.error( `Package "${ data.packageName }" is not available. Run "mgit bootstrap" in order to download the package.` );
+				log.error( `Package "${ data.packageName }" is not available. Run "mgit sync" in order to download the package.` );
 
 				return reject( { logs: log.all() } );
 			}
