+++ lib/commands/pull.js
@@ -20,7 +20,7 @@ module.exports = {
 
 		return `
     ${ u( 'Description:' ) }
-        Pull changes in all packages. If some package is missed, "${ i( 'bootstrap' ) }" command is calling on the missing package.
+        Pull changes in all packages. If some package is missed, the command will not be executed.
         For cloned repositories this command is a shorthand for: "${ i( 'mgit exec \'git pull\'' ) }".
 
     ${ u( 'Options:' ) }
@@ -36,30 +36,17 @@ module.exports = {
 	 */
 	execute( data ) {
 		const log = require( '../utils/log' )();
-		const bootstrapCommand = require( './bootstrap' );
 		const execCommand = require( './exec' );
 
 		const destinationPath = path.join( data.mgitOptions.packages, data.repository.directory );
 
 		// Package is not cloned.
 		if ( !fs.existsSync( destinationPath ) ) {
-			log.info( `Package "${ data.packageName }" was not found. Cloning...` );
+			log.info( `Package "${ data.packageName }" was not found. Skipping...` );
 
-			const bootstrapOptions = {
-				arguments: data.arguments,
-				mgitOptions: data.mgitOptions,
-				packageName: data.packageName,
-				repository: data.repository
-			};
-
-			return bootstrapCommand.execute( bootstrapOptions )
-				.then( response => {
-					log.concat( response.logs );
-
-					response.logs = log.all();
-
-					return Promise.resolve( response );
-				} );
+			return Promise.resolve( {
+				logs: log.all()
+			} );
 		}
 
 		return execCommand.execute( getExecData( 'git pull' ) );
