+++ lib/commands/sync.js
@@ -8,11 +8,13 @@
 const fs = require( 'fs' );
 const path = require( 'upath' );
 const chalk = require( 'chalk' );
+const shell = require( '../utils/shell' );
+const buildOptions = require( 'minimist-options' );
+const minimist = require( 'minimist' );
 
 module.exports = {
 	get helpMessage() {
 		const {
-			italic: i,
 			gray: g,
 			magenta: m,
 			underline: u
@@ -21,7 +23,7 @@ module.exports = {
 		return `
     ${ u( 'Description:' ) }
         Updates all packages. For packages that contain uncommitted changes, the update process is aborted. 
-        If some package is missed, "${ i( 'bootstrap' ) }" command is calling on the missing package.
+        If some package is missed, it will be installed automatically.
         
         The update process executes following commands:
 
@@ -32,7 +34,7 @@ module.exports = {
             * Pulls the changes if the repository is not detached at some commit.         
         
     ${ u( 'Options:' ) }
-        ${ m( '--recursive' ) }                 Whether to install dependencies recursively. Only packages matching these 
+        ${ m( '--recursive' ) } (-r)            Whether to install dependencies recursively. Only packages matching these 
                                     patterns will be cloned recursively.
                                     ${ g( 'Default: false' ) }
 		`;
@@ -44,30 +46,18 @@ module.exports = {
 	 */
 	execute( data ) {
 		const log = require( '../utils/log' )();
-		const bootstrapCommand = require( './bootstrap' );
 		const execCommand = require( './exec' );
 
 		const destinationPath = path.join( data.mgitOptions.packages, data.repository.directory );
 
 		// Package is not cloned.
 		if ( !fs.existsSync( destinationPath ) ) {
-			log.info( `Package "${ data.packageName }" was not found. Cloning...` );
-
-			const bootstrapOptions = {
-				arguments: data.arguments,
-				mgitOptions: data.mgitOptions,
-				packageName: data.packageName,
-				repository: data.repository
-			};
-
-			return bootstrapCommand.execute( bootstrapOptions )
-				.then( response => {
-					log.concat( response.logs );
-
-					response.logs = log.all();
-
-					return Promise.resolve( response );
-				} );
+			return this._clonePackage( {
+				path: destinationPath,
+				name: data.packageName,
+				url: data.repository.url,
+				branch: data.repository.branch
+			}, this._parseArguments( data.arguments ) );
 		}
 
 		return execCommand.execute( getExecData( 'git status -s' ) )
@@ -132,5 +122,63 @@ module.exports = {
 	 */
 	afterExecute( parsedPackages ) {
 		console.log( chalk.cyan( `${ parsedPackages.size } packages have been processed.` ) );
+	},
+
+	/**
+	 * @private
+	 * @param {Object} packageDetails
+	 * @param {Object} options Command options.
+	 * @returns {Promise}
+	 */
+	_clonePackage( packageDetails, options ) {
+		const log = require( '../utils/log' )();
+
+		log.info( `Package "${ packageDetails.name }" was not found. Cloning...` );
+
+		const command = [
+			`git clone --progress "${ packageDetails.url }" "${ packageDetails.path }"`,
+			`cd "${ packageDetails.path }"`,
+			`git checkout --quiet ${ packageDetails.branch }`
+		].join( ' && ' );
+
+		return shell( command )
+			.then( output => {
+				log.info( output );
+
+				const commandOutput = {
+					logs: log.all()
+				};
+
+				if ( options.recursive ) {
+					const packageJson = require( path.join( packageDetails.path, 'package.json' ) );
+					const packages = [];
+
+					if ( packageJson.dependencies ) {
+						packages.push( ...Object.keys( packageJson.dependencies ) );
+					}
+
+					if ( packageJson.devDependencies ) {
+						packages.push( ...Object.keys( packageJson.devDependencies ) );
+					}
+
+					commandOutput.packages = packages;
+				}
+
+				return commandOutput;
+			} );
+	},
+
+	/**
+	 * @private
+	 * @param {Array.<String>} argv List of arguments provided by the user via CLI.
+	 * @returns {Object}
+	 */
+	_parseArguments( argv ) {
+		return minimist( argv, buildOptions( {
+			recursive: {
+				type: 'boolean',
+				alias: 'r',
+			}
+		} ) );
 	}
 };
